pyproject.toml
[project]
name = "tg-gpt-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.25.0",
    "google-genai",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
]
